requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Django==4.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16